% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_support.R
\name{boldPlots}
\alias{boldPlots}
\title{Make text in ggplot2 objects bold, and set font family, size, and face}
\usage{
boldPlots(p = NULL)
}
\arguments{
\item{p}{A ggplot2 object created by a FASTAptameR visualization command}
}
\value{
The same ggplot2 object with updated aesthetics
}
\description{
Make text in ggplot2 objects bold, and set font family, size, and face
}
