% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDiversity_analyses.R
\name{fa_clusterDiversity}
\alias{fa_clusterDiversity}
\title{Computer cluster-specific metadata}
\usage{
fa_clusterDiversity(faDF_cluster = NULL)
}
\arguments{
\item{faDF_cluster}{A clustered data.frame generated by fa_clusterLED}
}
\value{
A data.frame that includes metadata for each cluster, including total unique sequences, total reads, total RPU, average LED, and Simpson's index of diversity
}
\description{
\code{fa_clusterDiversity} returns the following metadata per cluster: unique sequence count, total read count, total RPU, average LED, and Simpson's index of diversity.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, compute metadata
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(
  faDF = countData,
  minReads = 5,
  maxLED = 10,
  totalClusters = 100,
  keepNC = FALSE
)
cluster_metadata <- fa_clusterDiversity(faDF_cluster = clusterData)
}

}
\seealso{
\code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterDiversity_metaplot]{fa_clusterDiversity_metaplot()}}, \code{\link[=fa_clusterDiversity_kmer]{fa_clusterDiversity_kmer()}}
}
