% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDiversity_plots.R
\name{fa_clusterDiversity_kmer}
\alias{fa_clusterDiversity_kmer}
\title{Return a PCA or UMAP scatter plot from a k-mer matrix}
\usage{
fa_clusterDiversity_kmer(
  faDF_cluster = NULL,
  kmerSize = 3,
  clustersToPlot = c(1:3),
  plotType = "PCA",
  xaxis = "Dim1",
  yaxis = "Dim2",
  legend_title = "Cluster",
  plot_title = "Cluster k-mer plot",
  colour_palette_disc = "Dark2"
)
}
\arguments{
\item{faDF_cluster}{A clustered data.frame generated by fa_clusterLED}

\item{kmerSize}{Value of k used to generate k-mers (e.g., 3 generates 3-mers)}

\item{clustersToPlot}{Clusters to include in the visualization; clusters are named as integers}

\item{plotType}{One of \code{PCA} or \code{UMAP} (default: "PCA")}

\item{xaxis}{Name for the x-axis (default: "Dim1")}

\item{yaxis}{Name for the y-axis (default: "Dim2")}

\item{legend_title}{Title for the legend (default: "Cluster")}

\item{plot_title}{Title for the plot (default: "Cluster k-mer plot")}

\item{colour_palette_disc}{A discrete RColorBrewer palette (default: "Dark2")}
}
\value{
This function returns a PCA or UMAP plot based on kmer counts
}
\description{
\code{fa_clusterDiversity_kmer} returns a two-dimensional PCA or UMAP plot where each point corresponds to one sequence, and colors correspond to user-specified clusters (max = 8 clusters).
}
\details{
The function filters the input data.frame for 1-8 user-selected clusters and computes k-mer counts for each sequence
The final matrix has one row per sequence and one column per k-mer (e.g., AAA, AAC, etc. in the case of 3-mers), and this is the input to a PCA or UMAP.
Only the first two dimensions of PCA / UMAP are plotted, and points have colors corresponding to their respective clusters.
}
\examples{
\dontrun{
# count FASTQ file, cluster data
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)

fa_clusterDiversity_kmer(faDF_cluster = clusterData, kmerSize = 3, clustersToPlot = c(1:5))
}

}
\seealso{
\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}, \code{\link[=fa_clusterDiversity]{fa_clusterDiversity()}}, \code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterDiversity_metaplot]{fa_clusterDiversity_metaplot()}}
}
