% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDiversity_plots.R
\name{fa_clusterDiversity_metaplot}
\alias{fa_clusterDiversity_metaplot}
\title{Return line plots of cluster metadata}
\usage{
fa_clusterDiversity_metaplot(
  diversityDF = NULL,
  xaxis = "Cluster",
  yaxes = c("Seq. count", "Read count", "Avg. LED"),
  plot_title = "Cluster metaplots",
  line_colours = c("blue", "orange", "forestgreen")
)
}
\arguments{
\item{diversityDF}{A data.frame generated by \code{fa_clusterDiversity}}

\item{xaxis}{The name assigned to all x-axes in the plot; this axis corresponds to the cluster number (default: "Cluster")}

\item{yaxes}{The 3x y-axis names respectively assigned to the 3x subplots; these respectively correspond to number of unique sequences, number of total reads, and average LED (defaults: "Seq. count", "Read count", "Avg. LED")}

\item{plot_title}{The title of the plot (default: "Cluster metaplots")}

\item{line_colours}{The 3x colors respectively used for the 3x subplots (defaults: "blue", "orange", "forestgreen")}
}
\value{
This function returns 3x interactive plots of cluster metastats
}
\description{
\code{fa_clusterDiversity_metaplot} returns 3x subplots that show the unique sequence count, total read count, and average LED per cluster.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, compute metadata
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
diversityDF <- fa_clusterDiversity(faDF_cluster = clusterData)

# plot metadata
fa_clusterDiversity_metaplot(diversityDF = diversityDF)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_clusterDiversity]{fa_clusterDiversity()}}, \code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterDiversity_kmer]{fa_clusterDiversity_kmer()}}
}
