% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analyses.R
\name{fa_clusterLED}
\alias{fa_clusterLED}
\title{Cluster sequences with the Levenshtein edit distance (LED)}
\usage{
fa_clusterLED(
  faDF = NULL,
  minReads = 10,
  maxLED = 7,
  totalClusters = 30,
  keepNC = TRUE
)
}
\arguments{
\item{faDF}{A FASTAptameR data.frame that at least has columns for ID, Rank, Reads, RPU, and Sequence; this is generated with \code{fa_count} or many other functions in the package}

\item{minReads}{The minimum number of reads that a sequence needs to be considered for clustering (default: 10)}

\item{maxLED}{The maximum allowed Levenshtein Edit Distance (LED) allowed between an non-clustered sequence and a seed sequence (default: 7)}

\item{totalClusters}{The total number of clusters to generate; a sufficiently large number will result in the algorithm continuing until all sequences are clustered (default: 30)}

\item{keepNC}{A logical value that determines if non-clustered sequences should be kept (default: TRUE)}
}
\value{
A FASTAptameR data.frame that includes columns for Cluster, Rank in the Cluster (determined by RPU), and the LED between the sequence and the cluster's seed sequence
}
\description{
\code{fa_clusterLED} sorts input sequences according to their RPU (Reads Per Unit; computed in \code{fa_count}) and then omits sequences with a read count less than \code{minReads}.
The clustering algorithm proceeds according to the following steps:
\enumerate{
\item Set the most abundant, unclustered sequence as the cluster seed
\item Add all unclustered sequences with a user-defined LED to the cluster
\item Set aside clustered sequences and repeat
These steps repeat until the algorithm generates a user-defined number of clusters or until all sequences are clustered. Unclustered sequences are only kept if \code{keepNC} is TRUE.
}
}
\examples{
\dontrun{
# count FASTQ file, cluster data
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(
  faDF = countData,
  minReads = 5,
  maxLED = 10,
  totalClusters = 100,
  keepNC = FALSE
)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_clusterDiversity]{fa_clusterDiversity()}}, \code{\link[=fa_clusterMSA]{fa_clusterMSA()}}, \code{\link[=fa_recluster]{fa_recluster()}}
}
