% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSA_analyses.R
\name{fa_clusterMSA}
\alias{fa_clusterMSA}
\title{Perform an MSA for a single cluster of sequences}
\usage{
fa_clusterMSA(faDF_cluster = NULL, clusterSelection = 1, seq_type = "dna")
}
\arguments{
\item{faDF_cluster}{A clustered data.frame generated by fa_clusterLED}

\item{clusterSelection}{A single integer representing the cluster to be aligned (default: 1)}

\item{seq_type}{One of dna or protein (default: "dna")}
}
\value{
A single cluster in a data.frame that includes aligned sequences
}
\description{
\code{fa_clusterMSA} performs a DNA (default) or protein multiple sequence alignment (MSA) with the \code{msa} package.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, perform MSA
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
msaData <- fa_clusterMSA(faDF_cluster = clusterData, clusterSelection = 10, seq_type = "dna")
}

}
\seealso{
\code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterMSA_entropy]{fa_clusterMSA_entropy()}}, \code{\link[=fa_clusterMSA_mutInfo]{fa_clusterMSA_mutInfo()}}, \code{\link[msa:msa]{msa::msa()}}
}
