% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSA_plots.R
\name{fa_clusterMSA_entropy}
\alias{fa_clusterMSA_entropy}
\title{Return a bar chart showing the entropy per position of an alignment}
\usage{
fa_clusterMSA_entropy(
  msa_df = NULL,
  xaxis = "MSA position",
  yaxis = "Entropy (nats)",
  legend_title = "Nats",
  plot_title = "Entropy by position in MSA",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{msa_df}{A data.frame generated by \code{fa_clusterMSA}}

\item{xaxis}{The x-axis title (default: "MSA position")}

\item{yaxis}{The y-axis title (default: "Entropy (nats)")}

\item{legend_title}{The legend title (default: "Nats")}

\item{plot_title}{The plot title (default: "Entrpy by position in MSA")}

\item{bar_outline}{The color of the bar outlines (default: "black")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
A plotly object with a bar chart showing the entropy at every position of the alignment
}
\description{
\code{fa_clusterMSA_entropy} returns a bar chart that shows the entropy at every position of an aligned cluster.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, generate MSA
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
msaData <- fa_clusterMSA(faDF_cluster = clusterData)

# plot MSA entropy
fa_clusterMSA_entropy(msa_df = msa_df)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_clusterMSA]{fa_clusterMSA()}}, \code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterMSA_mutInfo]{fa_clusterMSA_mutInfo()}}
}
