% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSA_plots.R
\name{fa_clusterMSA_mutInfo}
\alias{fa_clusterMSA_mutInfo}
\title{Return a mutual information heat map}
\usage{
fa_clusterMSA_mutInfo(
  msa_df = NULL,
  xaxis = "MSA position",
  yaxis = "MSA position",
  legend_title = "MI",
  plot_title = "Pairwise mutual information in MSA",
  fill_palette_cont = "magma"
)
}
\arguments{
\item{msa_df}{A data.frame generated by \code{fa_clusterMSA}}

\item{xaxis}{The x-axis title (default: "MSA position")}

\item{yaxis}{The y-axis title (default: "MSA position")}

\item{legend_title}{The legend title (default: "MI")}

\item{plot_title}{The plot title (default: "Pairwise mutual information in MSA")}

\item{fill_palette_cont}{A continuous color palette (default: "magma")}
}
\value{
A plotly object with a heat map showing the mutual information between all pairs of positions in the alignment
}
\description{
\code{fa_clusterMSA_mutInfo} returns a heat map showing the mutual information between pairs of positions in aligned sequences.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, generate MSA
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
msaData <- fa_clusterMSA(faDF_cluster = clusterData)

# plot MSA mutual information
fa_clusterMSA_mutInfo(msa_df = msa_df)
}

}
\seealso{
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}}, \code{\link[=fa_clusterMSA]{fa_clusterMSA()}}, \code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterMSA_entropy]{fa_clusterMSA_entropy()}}
}
