% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plots.R
\name{fa_count_histogram}
\alias{fa_count_histogram}
\title{Return histograms of sequence lengths}
\usage{
fa_count_histogram(
  countData = NULL,
  xaxis = "Sequence length",
  yaxes = c("Unique sequences", "Read count"),
  plot_title = "Sequence-length histogram",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{countData}{A data.frame from \code{fa_count}}

\item{xaxis}{The title of the x-axis (default: "Sequence length")}

\item{yaxes}{The titles of the y-axis (defaults: "Unique sequences", "Read count")}

\item{plot_title}{The title of the plot (default: "Sequence-length histogram")}

\item{bar_outline}{The color for the bar outlines (default: "black")}

\item{bar_fill}{The color for the bar fills (default: "skyblue")}
}
\value{
A plotly object with two histograms for the sequence lengths for 1) unique reads and 2) all reads
}
\description{
\code{fa_count_histogram} returns two sequence-length histograms: 1) for unique sequences and 2) for all reads.
}
\examples{
\dontrun{
# count FASTQ files
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# make histogram
fa_count_histogram(countData = countData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_count_rpr]{fa_count_rpr()}}, \code{\link[=fa_count_binnedAbundance]{fa_count_binnedAbundance()}}
}
