% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_analyses.R
\name{fa_count_metadata}
\alias{fa_count_metadata}
\title{Compute metadata for a counted data.frame}
\usage{
fa_count_metadata(countData = NULL)
}
\arguments{
\item{countData}{A data.frame from fa_count}
}
\value{
A vector of metadata: total sequence count, unique sequence count, Simpson's index of diversity
}
\description{
\code{fa_count_metadata} prints a message containing the number of unique sequences, total number of reads, and the Simpson's index of diversity for a data.frame generated with \code{fa_count}.
}
\examples{
\dontrun{
# count FASTQ file, compute metadata
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
fa_count_metadata(countData = countData)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}
}
