% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plots.R
\name{fa_count_rpr}
\alias{fa_count_rpr}
\title{Return a reads-per-rank line plot}
\usage{
fa_count_rpr(
  countData = NULL,
  minReads = 10,
  maxRanks = 1000,
  xaxis = "Ranks of unique sequences",
  yaxis = "Total reads per unique sequence",
  plot_title = "Read count for each rank",
  line_color = "skyblue"
)
}
\arguments{
\item{countData}{A data.frame from \code{fa_count}}

\item{minReads}{The minimum number of reads required to include a sequence in the plot (default: 10)}

\item{maxRanks}{The maximum number of sequences to include (default: 1000)}

\item{xaxis}{The title of the x-axis (default: "Ranks of unique sequences")}

\item{yaxis}{The title of the y-axis (default: "Total reads per unique sequence")}

\item{plot_title}{The title of the full plot (default: "Read count for each rank")}

\item{line_color}{The color of the line (default: "skyblue")}
}
\value{
A plotly object with line plot showing the relationship between sequence rank and number of reads
}
\description{
\code{fa_count_rpr} returns a line plot showing the relationship between sequence rank and read count.
}
\examples{
\dontrun{
# count FASTQ files
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# make line plot
fa_count_rpr(countData = countData, minReads = 5, maxRanks = 1000)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_count_histogram]{fa_count_histogram()}}, \code{\link[=fa_count_binnedAbundance]{fa_count_binnedAbundance()}}
}
