% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMerge_analyses.R
\name{fa_dataMerge}
\alias{fa_dataMerge}
\title{Merge counted data.frames according to user-defined criteria}
\usage{
fa_dataMerge(faDF_list = NULL, mergeType = "full")
}
\arguments{
\item{faDF_list}{A list of data.frames generated with FASTAptameR}

\item{mergeType}{The type of merge to use when joining all data.frames in the list; "full" (default), "inner", or "left"}
}
\value{
A merged data.frame
}
\description{
\code{fa_dataMerge} merges a list of counted data.frames with a full, inner, or left join.
}
\examples{
\dontrun{
# count FASTQ files
countDataList <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq"),
  fa_count(dataInput = "PATH/TO/FILE3.fastq")
)

# full merge
fa_dataMerge(faDF_list = countDataList, mergeType = "full")
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_dataMerge_seqPersistence]{fa_dataMerge_seqPersistence()}}, \code{\link[=fa_dataMerge_UpSetR]{fa_dataMerge_UpSetR()}}
}
