% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMerge_plots.R
\name{fa_dataMerge_seqPersistence}
\alias{fa_dataMerge_seqPersistence}
\title{Return a sequence persistence bar chart}
\usage{
fa_dataMerge_seqPersistence(
  faDF_list = NULL,
  xaxis = "Population count",
  yaxis = "Unique read count",
  plot_title = "Sequence persistence plot",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{faDF_list}{A list of FASTAptameR data.frames}

\item{xaxis}{The title of the x-axis (default: "Population count")}

\item{yaxis}{The title of the y-axis (default: "Unique read count")}

\item{plot_title}{The title of the plot (default: "Sequence persistence plot")}

\item{bar_outline}{The color of the bar outlines (default: "black")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
A plotly object with a bar plot that shows the persistence of sequences across rounds
}
\description{
\code{fa_dataMerge_seqPersistence} returns a bar chart that shows the number of populations in which sequences are found (e.g., 100 sequences found in 3 populations, 40 sequences found in 2 populations, etc.).
}
\examples{
\dontrun{
# count FASTQ files
countDataList <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq"),
  fa_count(dataInput = "PATH/TO/FILE3.fastq")
)

# make persistence plot
fa_dataMerge_seqPersistence(faDF_list = countDataList)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_dataMerge]{fa_dataMerge()}}, \code{\link[=fa_dataMerge_UpSetR]{fa_dataMerge_UpSetR()}}
}
