% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMerge_plots.R
\name{fa_dataMerge_UpSetR}
\alias{fa_dataMerge_UpSetR}
\title{Return an UpSetR plot}
\usage{
fa_dataMerge_UpSetR(
  faDF_list = NULL,
  fastaNames = NULL,
  xaxis = "Sequences per set",
  yaxis = "Sequence intersections",
  bar_fill = "skyblue"
)
}
\arguments{
\item{faDF_list}{A list of FASTAptameR data.frames}

\item{fastaNames}{A list of names to use in the plot; should be one name per data.frame in faDF_list}

\item{xaxis}{The title of the x-axis (default: "Sequences per set")}

\item{yaxis}{The title of the y-axis (default: "Sequence intersections")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
An UpSet plot
}
\description{
\code{fa_dataMerge_UpSetR} returns an UpSetR plot showing the sequences shared across multiple populations
}
\examples{
\dontrun{
# count FASTQ files
countDataList <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq"),
  fa_count(dataInput = "PATH/TO/FILE3.fastq")
)

# make UpSet plot
fa_dataMerge_UpSetR(faDF_list = countDataList)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_dataMerge]{fa_dataMerge()}}, \code{\link[=fa_dataMerge_seqPersistence]{fa_dataMerge_seqPersistence()}}
}
