% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_analyses.R
\name{fa_distance}
\alias{fa_distance}
\title{Compute the LED between a query sequence and every other sequence in the target file}
\usage{
fa_distance(faDF = NULL, querySequence = NULL)
}
\arguments{
\item{faDF}{A FASTAptameR data.frame}

\item{querySequence}{A query sequence used as a reference for computing the Levenshtein edit distance (LED)}
}
\value{
A data.frame with a column for the LED between the sequence and query sequence
}
\description{
\code{fa_distance} computes the LED between a user-defined query sequence and every other sequence in the input data.frame.
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# compute distance
fa_distance(faDF = countData, querySequence = query)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_distance_histogram]{fa_distance_histogram()}}
}
