% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_plots.R
\name{fa_distance_histogram}
\alias{fa_distance_histogram}
\title{Return distance histograms}
\usage{
fa_distance_histogram(
  distanceData = NULL,
  xaxis = "Distance from query",
  yaxes = c("Unique sequences", "Read count"),
  plot_title = "Distance histograms",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{distanceData}{A data.frame from \code{fa_distance}}

\item{xaxis}{The title of the x-axis (default: "Distance from query")}

\item{yaxes}{The titles of the y-axis (defaults: "Unique sequences", "Read count")}

\item{plot_title}{The title of the plot (default: "Distance histograms")}

\item{bar_outline}{The color of the bar outlines (default: "black")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
A plotly object with a histogram of distances between all sequences and a query sequence
}
\description{
\code{fa_distance_histogram} returns two histograms (1. for unique sequences and 2. for all reads) to visualize the Levenshtein edit distance (LED) between a query sequence and all other sequences in a population
}
\examples{
\dontrun{
# count FASTQ file, compute LED between counted sequences and query sequence
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
distanceData <- fa_distance(faDF = countData, querySequence = query)

# make histogram
fa_distance_histogram(distanceData = distanceData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_distance]{fa_distance()}}
}
