% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_analyses.R
\name{fa_edgeR_pairTest}
\alias{fa_edgeR_pairTest}
\title{Perform edgeR-based differential analysis between two sets of conditions}
\usage{
fa_edgeR_pairTest(faDF_cond1 = NULL, faDF_cond2 = NULL, pcutoff = 0.1)
}
\arguments{
\item{faDF_cond1}{A list of FASTAptameR data.frames for the first condition}

\item{faDF_cond2}{A list of FASTAptameR data.frames for the second condition}

\item{pcutoff}{A numeric threshold for the BH-corrected p-values; used to determine significance (default: 0.1)}
}
\value{
A data.frame with the results from the edgeR test
}
\description{
\code{fa_edgeR_pairTest} performs a pairwise differential analysis test between two conditions.
Each condition should be represented as a list of counted data.frames.
}
\examples{
\dontrun{
# count FASTQ files in condition 1
countData_cond1 <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq"),
  fa_count(dataInput = "PATH/TO/FILE3.fastq")
)

# count FASTQ files in condition 2
countData_cond2 <- list(
  fa_count(dataInput = "PATH/TO/FILE4.fastq"),
  fa_count(dataInput = "PATH/TO/FILE5.fastq"),
  fa_count(dataInput = "PATH/TO/FILE6.fastq")
)

# use edgeR
edgeRData <- fa_edgeR_pairTest(
  faDF_cond1 = countData_cond1,
  faDF_cond2 = countData_cond2,
  pcutoff = 0.05
)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_edgeR_pairTestViz]{fa_edgeR_pairTestViz()}}, \code{\link[edgeR:exactTest]{edgeR::exactTest()}}
}
