% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_plots.R
\name{fa_edgeR_pairTestViz}
\alias{fa_edgeR_pairTestViz}
\title{Return a scatter plot with the results of a paired DEA test with edgeR}
\usage{
fa_edgeR_pairTestViz(
  edgeR_DF = NULL,
  xaxis = "logCPM",
  yaxis = "logFC",
  plot_title = "edgeR results",
  point_colours = c("red", "black")
)
}
\arguments{
\item{edgeR_DF}{A data.frame from \code{fa_edgeR}}

\item{xaxis}{The x-axis title (default: "logCPM")}

\item{yaxis}{The y-axis title (default: "logFC")}

\item{plot_title}{The plot title (default: "edgeR results")}

\item{point_colours}{A vector of color names for significant (default: "red") and insignificant (default: "black") points}
}
\value{
A plotly object with a scatter plot of logCPM vs logFC from the edgeR table
}
\description{
\code{fa_edgeR_pairTestViz} returns a scatter plot of logCPM vs logFC.
Point color and shape respectively correspond to whether the sequence is considered significant based on corrected p-values from edgeR and a user-defined threshold in \code{fa_edgeR()}.
}
\examples{
\dontrun{
# count FASTQ files in condition 1
countData_cond1 <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq"),
  fa_count(dataInput = "PATH/TO/FILE3.fastq")
)

# count FASTQ files in condition 2
countData_cond2 <- list(
  fa_count(dataInput = "PATH/TO/FILE4.fastq"),
  fa_count(dataInput = "PATH/TO/FILE5.fastq"),
  fa_count(dataInput = "PATH/TO/FILE6.fastq")
)

# use edgeR
edgeRData <- fa_edgeR_pairTest(
  faDF_cond1 = countData_cond1,
  faDF_cond2 = countData_cond2,
  pcutoff = 0.05
)

# make volcano plot
fa_edgeR_pairTestViz(edgeR_DF = edgeRData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_edgeR_pairTest]{fa_edgeR_pairTest()}}
}
