% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqEnrich_analyses.R
\name{fa_enrich}
\alias{fa_enrich}
\title{Compute the enrichment of sequences from multiple populations}
\usage{
fa_enrich(faDF1 = NULL, faDF2 = NULL, keepNA = FALSE)
}
\arguments{
\item{faDF1}{A FASTAptameR data.frame}

\item{faDF2}{A FASTAptameR data.frame}

\item{keepNA}{Logical indicating if sequences only present in a single population should be kept (default: FALSE)}
}
\value{
A merged data.frame with columns for enrichment and log2(enrichment)
}
\description{
\code{fa_enrich} merges two counted data.frames by sequence and computes sequence enrichment by considering the ratio of the RPU values.
Thus, it is critical to use the same scaling factor when normalizing the read counts in \code{fa_count}.
}
\examples{
\dontrun{
# count FASTQ files
countDataList <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)

# compute sequence enrichment between the two files
enrichData <- fa_enrich(faDF1 = countDataList[[1]], faDF2 = countDataList[[2]])
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_enrich_histogram]{fa_enrich_histogram()}} \code{\link[=fa_enrich_scatter]{fa_enrich_scatter()}}, \code{\link[=fa_enrich_ra]{fa_enrich_ra()}}
}
