% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqEnrich_plots.R
\name{fa_enrich_histogram}
\alias{fa_enrich_histogram}
\title{Return a histogram of fold changes}
\usage{
fa_enrich_histogram(
  enrichDF = NULL,
  xaxis = "log2(Enrichment)",
  yaxis = "Unique sequence count",
  plot_title = "log2E histogram",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{enrichDF}{An enrichment data.frame from \code{fa_enrich}}

\item{xaxis}{The title of the x-axis (default: "log2(Enrichment)")}

\item{yaxis}{The title of the y-axis (default: "Unique sequence count")}

\item{plot_title}{The title of the plot (default: "log2E histogram")}

\item{bar_outline}{The color of the bar outlines (default: "black")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
A plotly object with a histogram of enrichment scores
}
\description{
\code{fa_enrich_histogram} returns a histogram of sequence enrichment scores after a log2 transformation.
}
\examples{
\dontrun{
# count FASTQ files, compute enrichment
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
enrichData <- fa_enrich(faDF1 = countData[[1]], faDF2 = countData[[2]])

# make histogram
fa_enrich_histogram(enrichDF = enrichData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_enrich]{fa_enrich()}}, \code{\link[=fa_enrich_scatter]{fa_enrich_scatter()}}, \code{\link[=fa_enrich_ra]{fa_enrich_ra()}}
}
