% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqEnrich_plots.R
\name{fa_enrich_ra}
\alias{fa_enrich_ra}
\title{Return an RA plot from RPUs in two populations}
\usage{
fa_enrich_ra(
  enrichDF = NULL,
  xaxis = "Average log2(RPU)",
  yaxis = "Fold change",
  plot_title = "Enrichment RA plot",
  point_colour = "skyblue"
)
}
\arguments{
\item{enrichDF}{An enrichment data.frame from \code{fa_enrich}}

\item{xaxis}{The title of the x-axis (default: "Average log2(RPU)")}

\item{yaxis}{The title of the y-axis (default: "Fold change")}

\item{plot_title}{The title of the plot (default: "Enrichment RA plot")}

\item{point_colour}{The color of the points (default: "skyblue")}
}
\value{
A plotly object with an RA scatter plot
}
\description{
\code{fa_enrich_ra} generates a ratio average (RA) plot from RPU values in the 1st and 2nd populations (RPU.a and RPU.b, respectively).
A small epsilon factor (epsilon = 0.001) is added to the RPU values in both populations to avoid dividing by 0s or taking the log2 of them.
R of the RA plot is defined as log2(RPU.b / RPU.a), which corresponds to the fold change in RPU.
A of the RA plot is defined as log2(RPU.b * RPU.a) / 2, which corresponds to the average log2-transformed RPU.
}
\examples{
\dontrun{
# count FASTQ files, compute enrichment
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
enrichData <- fa_enrich(faDF1 = countData[[1]], faDF2 = countData[[2]])

# make histogram
fa_enrich_ra(enrichDF = enrichData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_enrich]{fa_enrich()}}, \code{\link[=fa_enrich_histogram]{fa_enrich_histogram()}}, \code{\link[=fa_enrich_scatter]{fa_enrich_scatter()}}
}
