% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqEnrich_plots.R
\name{fa_enrich_scatter}
\alias{fa_enrich_scatter}
\title{Return a scatter plot of RPUs in two populations}
\usage{
fa_enrich_scatter(
  enrichDF = NULL,
  xaxis = "RPU.a",
  yaxis = "RPU.b",
  plot_title = "RPU.a vs RPU.b",
  point_colour = "skyblue"
)
}
\arguments{
\item{enrichDF}{An enrichment data.frame from \code{fa_enrich}}

\item{xaxis}{The title of the x-axis (default: "RPU.a")}

\item{yaxis}{The title of the y-axis (default: "RPU.b")}

\item{plot_title}{The title of the plot (default: "RPU.a vs RPU.b")}

\item{point_colour}{The color of the points (default: "skyblue")}
}
\value{
A plotly object with a scatter plot RPU in the first population vs RPU in the second population
}
\description{
\code{fa_enrich_scatter} returns a scatter plot in which points correspond to sequences and coordinates correspond to RPU in the 1st and 2nd populations (RPU.a and RPU.b, respectively).
}
\examples{
\dontrun{
# count FASTQ files, compute enrichment
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
enrichData <- fa_enrich(faDF1 = countData[[1]], faDF2 = countData[[2]])

# make histogram
fa_enrich_scatter(enrichDF = enrichData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_enrich]{fa_enrich()}}, \code{\link[=fa_enrich_histogram]{fa_enrich_histogram()}}, \code{\link[=fa_enrich_ra]{fa_enrich_ra()}}
}
