% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_support.R
\name{fa_formatInput}
\alias{fa_formatInput}
\title{Read a FASTAptameR-formatted FASTA file into a data.frame}
\usage{
fa_formatInput(fastaInput = NULL)
}
\arguments{
\item{fastaInput}{A FASTA file that is in the FASTAptameR format: sequence identifiers and sequences on alternating lines; the identifier must contain at least the rank, read count, and RPU in a hyphen-separated string}
}
\value{
A data.frame that contains the same information from the FASTAptameR FASTA file
}
\description{
Read a FASTAptameR-formatted FASTA file into a data.frame
}
\examples{
\dontrun{
# read into data.frame
countData <- fa_formatInput(fastaInput = "PATH/TO/FASTAPTAMER.fasta")
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_formatOutput]{fa_formatOutput()}}
}
