% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_support.R
\name{fa_formatOutput}
\alias{fa_formatOutput}
\title{Format a FASTAptameR data.frame as a FASTA file}
\usage{
fa_formatOutput(outputData = NULL)
}
\arguments{
\item{outputData}{A FASTAptameR data.frame like the one generated from fa_formatInput}
}
\value{
A FASTA-like vector with the same information as the FASTAptameR data.frame
}
\description{
Format a FASTAptameR data.frame as a FASTA file
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# format as FASTA, and write to file
countFASTA <- fa_formatOutput(outputData = countData)
writeLines(countFASTA, "PATH/TO/OUTPUT.fasta")
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_formatInput]{fa_formatInput()}}
}
