% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDiscovery_analyses.R
\name{fa_fsbc_motifDiscovery}
\alias{fa_fsbc_motifDiscovery}
\title{Identify over-enriched motifs}
\usage{
fa_fsbc_motifDiscovery(faDF = NULL, minReads = 10, lengthRange = c(5, 10))
}
\arguments{
\item{faDF}{A FASTAptameR data.frame}

\item{minReads}{The minimum read count for a sequence to be included in the analysis (default: 10)}

\item{lengthRange}{A numeric vector with the minimum and maximum length of motifs to consider (default: 5, 10)}
}
\value{
A data.frame with detected motifs
}
\description{
\code{fa_fsbc_motifDiscovery} is a wrapper around the FSBC code, which identifies over-enriched substrings from an input sequence set.
This code is adapted from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# find overrepresented motifs with the FSBC tools
motifData <- fa_fsbc_motifDiscovery(faDF = countData)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_fsbc_rootMotifs]{fa_fsbc_rootMotifs()}}, \code{\link[=fa_fsbc_motifDiscoveryPlot]{fa_fsbc_motifDiscoveryPlot()}}
}
