% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDiscovery_plots.R
\name{fa_fsbc_motifDiscoveryPlot}
\alias{fa_fsbc_motifDiscoveryPlot}
\title{Return a scatter plot with the results of the motif discovery module}
\usage{
fa_fsbc_motifDiscoveryPlot(
  motif_discovery_df = NULL,
  xaxis = "Rank by normalized z-score",
  yaxis = "-log10(p)",
  legend_title = "Length",
  plot_title = "Over-enriched strings",
  colour_palette_cont = "magma"
)
}
\arguments{
\item{motif_discovery_df}{A data.frame from \code{fa_motifDiscovery}}

\item{xaxis}{The title of the x-axis (default: "Rank by normalized z-score")}

\item{yaxis}{The title of the y-axis (default: "-log10(p)")}

\item{legend_title}{The title of the legend (default: "Length")}

\item{plot_title}{The title of the plot (default: "Over-enriched strings")}

\item{colour_palette_cont}{A continuous color palette (default: "magma")}
}
\value{
A plotly object with a bubble plot showing Rank vs p-value for detected motifs
}
\description{
\code{fa_fsbc_motifDiscoveryPlot} returns a scatter plot that displays the results of the motif discovery module (\code{fa_fsbc_motifDiscovery} and \code{fa_motifDiscovery}), which is built with the FSBC tool (DOI: 10.1186/s12859-020-03607-1).
The x-axis, y-axis, and color respectively correspond to the motif's 1) rank after sorting by normalized z-scores, 2) -log10(p-value), and 3) length.
}
\examples{
\dontrun{
# count FASTQ file, find overrepresented motifs with the FSBC tools
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
motifData <- fa_fsbc_rootMotifs(motif_discovery_df = fa_fsbc_motifDiscovery(faDF = countData))

# make scatter plot
fa_fsbc_motifDiscoveryPlot(motif_discover_df = motifData)
}

}
\seealso{
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_colour_viridis_c()}}, \code{\link[=fa_fsbc_motifDiscovery]{fa_fsbc_motifDiscovery()}}
}
