% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDiscovery_analyses.R
\name{fa_fsbc_rootMotifs}
\alias{fa_fsbc_rootMotifs}
\title{Collapse shorter sub-motifs into larger root motifs}
\usage{
fa_fsbc_rootMotifs(motif_discovery_df = NULL)
}
\arguments{
\item{motif_discovery_df}{A data.frame from fa_fsbc_motifDiscovery}
}
\value{
A condensed data.frame in which shorter motifs that are perfectly matched to longer motifs are merged
}
\description{
\code{fa_fsbc_rootMotifs} condenses the output of \code{fa_fsbc_motifDiscovery}, which may return sub-motifs.
We define "sub-motifs" as a shorter motif that can be found in its entirety in a longer motif (e.g., "AACG" is a sub-motif of "AACGTTT").
In this case, the count of occurrences for the sub-motif will be added to the longer motif.
However, one sub-motif may map back onto multiple longer motifs (e.g., "AACG" maps onto "AACGTTT" and "AACGTTA").
In this example, the sub-motif count will be added to both of the longer motifs.
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# find motifs with FSBC and condense motifs
motifData <- fa_fsbc_motifDiscovery(faDF = countData) \%>\%
  fa_fsbc_rootMotifs(motif_discovery_df = .)
}

}
\seealso{
@seealso \code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_fsbc_motifDiscovery]{fa_fsbc_motifDiscovery()}}, \code{\link[=fa_fsbc_motifDiscoveryPlot]{fa_fsbc_motifDiscoveryPlot()}}
}
