% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMSA_analyses.R
\name{fa_getClusters}
\alias{fa_getClusters}
\title{This function returns the clusters available in the input data}
\usage{
fa_getClusters(clusterPath = NULL)
}
\arguments{
\item{clusterPath}{The file path to a clustered FASTA file}
}
\value{
A vector of clusters found in the file
}
\description{
This function returns the clusters available in the input data
}
