% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_support.R
\name{fa_motif_format}
\alias{fa_motif_format}
\title{Format a motif query list as a Perl-like regular expression}
\usage{
fa_motif_format(motifList = NULL, motifType = "Nucleotide")
}
\arguments{
\item{motifList}{A comma-separated list of motifs (i.e., a vector of length 1)}

\item{motifType}{One of Nucleotide (default) or Protein; when Nucleotide is selected, the script will interpret nucleotide ambiguity codes}
}
\value{
A formatted list of motifs
}
\description{
Format a motif query list as a Perl-like regular expression
}
