% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifTracker_analyses.R
\name{fa_motif_motifTracker}
\alias{fa_motif_motifTracker}
\title{Track motif occurrences across populations}
\usage{
fa_motif_motifTracker(
  faDF_list = NULL,
  populationNames = NULL,
  queryList = NULL,
  queryAliases = NULL,
  motifType = "Nucleotide"
)
}
\arguments{
\item{faDF_list}{A list of FASTAptameR data.frames; there should be one data.frame per population name}

\item{populationNames}{A list of names for each data.frame; there should be one population name per data.frame}

\item{queryList}{A list of motifs to track across populations}

\item{queryAliases}{An optional list of aliases for the motifs; if supplied, there must be one alias per motif}

\item{motifType}{One of Nucleotide (default) or Protein; the Nucleotide option considers nucleotide ambiguity codes}
}
\value{
A data.frame with data related to motif occurrence(s) across multiple populations
}
\description{
\code{fa_motif_motifTracker} tracks the occurrences of motifs across multiple input populations.
}
\examples{
\dontrun{
# count FASTQ files
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)

# count motifs in input files
trackerData <- fa_motif_motifTracker(
  faDF_list = countData,
  populationNames = c("Population A", "Population B"),
  queryList = c("AAA", "AACG"),
  queryAliases = c("motif1", "motif2"),
  motifType = "Nucleotide"
)

# compute motif enrichment
enrichData <- fa_motif_trackerEnrichment(trackerDF = trackerData)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_motif_sequenceTracker]{fa_motif_sequenceTracker()}}, \code{\link[=fa_motif_trackerEnrichment]{fa_motif_trackerEnrichment()}}, \code{\link[=fa_motif_motifTrackerPlot]{fa_motif_motifTrackerPlot()}}
}
