% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifTracker_plots.R
\name{fa_motif_motifTrackerPlot}
\alias{fa_motif_motifTrackerPlot}
\title{Return a line plot of motif abundance across populations}
\usage{
fa_motif_motifTrackerPlot(
  trackerDF = NULL,
  xaxis = "Population",
  yaxis = "Percentage",
  plot_title = "Motif tracker",
  colour_palette_disc = "Dark2"
)
}
\arguments{
\item{trackerDF}{A data.frame from \code{fa_motif_motifTracker}}

\item{xaxis}{The title of the x-axis (default: "Population")}

\item{yaxis}{The title of the y-axis (default: "Percentage")}

\item{plot_title}{The title of the plot (default: "Motif tracker")}

\item{colour_palette_disc}{A discrete RColorBrewer palette (default: "Dark2")}
}
\value{
A plotly object with a line plot showing motif occurrence in each population
}
\description{
\code{fa_motif_motifTrackerPlot} returns a line plot that shows the percentage of sequences with user-defined motifs across multiple populations.
This current formulation assumes that populations have some form of temporal dependence, such as consecutive rounds in a SELEX experiment.
}
\examples{
\dontrun{
# count FASTQ files, count motifs across files
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
trackerData <- fa_motif_motifTracker(
  faDF_list = countData,
  populationNames = c("Population A", "Population B"),
  queryList = c("AAA", "AACG"),
  queryAliases = c("motif1", "motif2"),
  motifType = "Nucleotide"
)

# make line plot
fa_motif_motifTrackerPlot(trackerDF = trackerData)
}

}
\seealso{
\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}, \code{\link[=fa_motif_motifTracker]{fa_motif_motifTracker()}}
}
