% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifTracker_analyses.R
\name{fa_motif_sequenceTracker}
\alias{fa_motif_sequenceTracker}
\title{Track sequence occurrence across populations}
\usage{
fa_motif_sequenceTracker(
  faDF_list = NULL,
  populationNames = NULL,
  queryList = NULL,
  queryAliases = NULL
)
}
\arguments{
\item{faDF_list}{A list of FASTAptameR data.frames; there should be one data.frame per population name}

\item{populationNames}{A list of names for each data.frame; there should be one population name per data.frame}

\item{queryList}{A list of sequences to track across populations}

\item{queryAliases}{An optional list of aliases for the motifs; if supplied, there must be one alias per motif}
}
\value{
A data.frame with data related to sequence occurrence(s) across multiple populations
}
\description{
\code{fa_motif_sequenceTracker} tracks the occurrences of sequences across multiple input populations.
}
\examples{
\dontrun{
# count FASTQ files
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)

# count sequence occurrences across files
trackerData <- fa_motif_sequenceTracker(
  faDF_list = countData,
  populationNames = c("Population A", "Population B"),
  queryList = c(sequence1, sequence2),
  queryAliases = c("seq1", "seq2")
)

# compute sequence enrichment
enrichData <- fa_motif_trackerEnrichment(trackerDF = trackerData)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_motif_motifTracker]{fa_motif_motifTracker()}}, \code{\link[=fa_motif_trackerEnrichment]{fa_motif_trackerEnrichment()}}, \code{\link[=fa_motif_sequenceTrackerPlot]{fa_motif_sequenceTrackerPlot()}}
}
