% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifTracker_plots.R
\name{fa_motif_sequenceTrackerPlot}
\alias{fa_motif_sequenceTrackerPlot}
\title{Return a line plot of sequence abundance across populations}
\usage{
fa_motif_sequenceTrackerPlot(
  trackerDF = NULL,
  xaxis = "Population",
  yaxis = "RPU",
  plot_title = "Sequence tracker",
  colour_palette_disc = "Dark2"
)
}
\arguments{
\item{trackerDF}{A data.frame from \code{fa_motif_sequenceTracker}}

\item{xaxis}{The x-axis title (default: "Population")}

\item{yaxis}{The y-axis title (default: "RPU")}

\item{plot_title}{The plot title (default: "Sequence tracker")}

\item{colour_palette_disc}{A discrete RColorBrewer palette (default: "Dark2")}
}
\value{
A plotly object with a line plot showing sequence occurrence in each population
}
\description{
\code{fa_motif_sequenceTrackerPlot} returns a line plot showing the RPU of user-defined sequences across multiple populations.
The current formulation assumes that populations have some form of temporal dependence, such as consecutive rounds in a SELEX experiment.
}
\examples{
\dontrun{
# count FASTQ files, count sequence occurrences across files
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
trackerData <- fa_motif_sequenceTracker(
  faDF_list = countData,
  populationNames = c("Population A", "Population B"),
  queryList = c(sequence1, sequence2),
  queryAliases = c("seq1", "seq2")
)

# make line plot
fa_motif_sequenceTrackerPlot(trackerDF = trackerData)
}

}
\seealso{
\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}, \code{\link[=fa_motif_sequenceTracker]{fa_motif_sequenceTracker()}}
}
