% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifOmit_analyses.R
\name{fa_motifOmit}
\alias{fa_motifOmit}
\title{Omit sequences with user-defined motifs}
\usage{
fa_motifOmit(
  faDF = NULL,
  motif = NULL,
  partial = FALSE,
  motifType = "Nucleotide"
)
}
\arguments{
\item{faDF}{A FASTAptameR data.frame}

\item{motif}{A comma-separated list of motifs (e.g., "ACT,AAA,ACG")}

\item{partial}{A logical indicating whether partial matches are allowed (default: FALSE)}

\item{motifType}{A character string indicating the type of motif (default: "Nucleotide"); Nucleotide motifs allow ambiguities}
}
\value{
A filtered data.frame that omits sequences with the supplied motifs
}
\description{
\code{fa_motifOmit} omits sequences from a counted data.frame if they have at least one user-defined motif (\code{partial} = TRUE) or all user-defined motifs (\code{partial} = FALSE)
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# omit sequences with a user-defined motif
countData_omit <- fa_motifOmit(
  faDF = countData,
  motif = "UUCG,CGGG",
  partial = FALSE,
  motifType = "Nucleotide"
)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_motifSearch]{fa_motifSearch()}}
}
