% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifSearch_analyses.R
\name{fa_motifSearch}
\alias{fa_motifSearch}
\title{Keep sequences that have user-defined motifss}
\usage{
fa_motifSearch(
  faDF = NULL,
  motif = NULL,
  highlight = FALSE,
  partial = FALSE,
  motifType = "Nucleotide"
)
}
\arguments{
\item{faDF}{A FASTAptameR data.frame}

\item{motif}{A comma-separated list of motifs (e.g., "ACT,AAA,ACG")}

\item{highlight}{A logical indicating whether motifs should be placed inside of parentheses (default: FALSE)}

\item{partial}{A logical indicating whether partial matches are allowed (default: FALSE)}

\item{motifType}{A character string indicating the type of motif (default: "Nucleotide"); Nucleotide motifs allow ambiguities}
}
\value{
A filtered data.frame that only includes sequences with the supplied motifs
}
\description{
\code{fa_motifSearch} only keeps sequences from a counted data.frame if they have at least one user-defined motif (\code{partial} = TRUE) or all user-defined motifs (\code{partial} = FALSE)
}
\examples{
\dontrun{
# count FASTQ file
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")

# only keep sequences with a user-defined motif
countData_search <- fa_motifSearch(
  faDF = countData,
  motif = "UUCG,CGGG",
  partial = FALSE,
  motifType = "Nucleotide"
)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[=fa_motifOmit]{fa_motifOmit()}}
}
