% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPHMM_analyses.R
\name{fa_phmm_create}
\alias{fa_phmm_create}
\title{Derive a PHMM from the MSA of a single cluster}
\usage{
fa_phmm_create(faDF_msa = NULL)
}
\arguments{
\item{faDF_msa}{A data.frame with aligned sequences from fa_clusterMSA}
}
\value{
An object of the "PHMM" class from the \code{aphid} package
}
\description{
\code{fa_phmm_create} derives a profile hidden Markov model (PHMM) for a single aligned cluster.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, perform MSA, derive PHMM
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
msaData <- fa_clusterMSA(faDF_cluster = clusterData, clusterSelection = 10, seq_type = "dna")
phmmObj <- fa_phmm_create(faDF_msa = msaData)
}

}
\seealso{
\code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterMSA]{fa_clusterMSA()}}, \code{\link[=fa_phmm_simulate]{fa_phmm_simulate()}}, \code{\link[aphid:derivePHMM]{aphid::derivePHMM()}}
}
