% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPHMM_analyses.R
\name{fa_phmm_simulate}
\alias{fa_phmm_simulate}
\title{Simulate sequences according to a derived PHMM}
\usage{
fa_phmm_simulate(msaPHMM = NULL, numSeq = 10, seqLength = 70)
}
\arguments{
\item{msaPHMM}{An object of the "PHMM" class; generated with fa_phmm_create}

\item{numSeq}{The number of sequences to simulate (default: 10)}

\item{seqLength}{The target length of simulated sequences (default: 70)}
}
\value{
A vector of simulated sequences
}
\description{
\code{fa_phmm_simulate} simulates sequences from a PHMM derived from \code{fa_phmm_create}.
}
\examples{
\dontrun{
# count FASTQ file, cluster data, perform MSA, derive PHMM, generate sequences with PHMM
countData <- fa_count(dataInput = "PATH/TO/FILE.fastq")
clusterData <- fa_clusterLED(faDF = countData)
msaData <- fa_clusterMSA(faDF_cluster = clusterData, clusterSelection = 10, seq_type = "dna")
phmmObj <- fa_phmm_create(faDF_msa = msaData)
sim_seqs <- fa_phmm_simulate(msaPHMM = phmmObj, numSeq = 20, seqLength = 56)
}

}
\seealso{
\code{\link[=fa_clusterLED]{fa_clusterLED()}}, \code{\link[=fa_clusterMSA]{fa_clusterMSA()}}, \code{\link[=fa_phmm_create]{fa_phmm_create()}}, \code{\link[aphid:generate]{aphid::generate()}}
}
