% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posEnrich_analyses.R
\name{fa_posEnrich}
\alias{fa_posEnrich}
\title{Compute the average enrichment per position per residue for a single super-cluster}
\usage{
fa_posEnrich(faDF_recluster = NULL, cluster_selection = 1, seq_type = "dna")
}
\arguments{
\item{faDF_recluster}{A reclustered data.frame from fa_recluster}

\item{cluster_selection}{The cluster to be considered for this analysis (default: 1); must be an integer that corresponds to a cluster present in the data}

\item{seq_type}{One of dna (default) or protein}
}
\value{
A data.frame with the average enrichment value for each residue at each position
}
\description{
\code{fa_posEnrich} computes the average enrichment of reclustered, aligned sequences for every residue at every position for a single super-cluster.
This function first aligns all sequences in a user-defined super-cluster.
Compute the average sequence enrichment for each column of the MSA and each valid character in the MSA (including the gap character, "-").
For a DNA example, compute the average enrichment for sequences with (A,C,G,T,-) at position 1, and then repeat this for every subsequent position.
Thus, this example will have 5 x L average enrichment scores, where 5 corresponds to the alphabet size, and L corresponds to the length of the aligned sequences.
}
\examples{
\dontrun{
# count FASTQ files, cluster data, recluster populations together
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
clusterData <- list(
  fa_clusterLED(faDF = countData[[1]]),
  fa_clusterLED(faDF = countData[[2]])
)
reclusterData <- fa_recluster(
  faDF1_cluster = clusterData[[1]],
  faDF2_cluster = clusterData[[2]]
)

# compute position-level enrichment for the 3rd super-cluster
posEnrichData <- fa_posEnrich(
  faDF_recluster = reclusterData,
  cluster_selection = 3,
  seq_type = "dna"
)
}

}
\seealso{
\code{\link[=fa_recluster]{fa_recluster()}}, \code{\link[msa:msa]{msa::msa()}}, \code{\link[=fa_posEnrich_barplot]{fa_posEnrich_barplot()}}, \code{\link[=fa_posEnrich_heatmap]{fa_posEnrich_heatmap()}}
}
