% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posEnrich_plots.R
\name{fa_posEnrich_barplot}
\alias{fa_posEnrich_barplot}
\title{Return a bar plot with the average enrichment per position, regardless of the character at that position}
\usage{
fa_posEnrich_barplot(
  faDF_posEnrich = NULL,
  xaxis = "Position",
  yaxis = "Avg. enrichment",
  plot_title = "Enrichment per position",
  bar_outline = "black",
  bar_fill = "skyblue"
)
}
\arguments{
\item{faDF_posEnrich}{A data.frame from \code{fa_posEnrich}}

\item{xaxis}{The title of the x-axis (default: "Position")}

\item{yaxis}{The title of the y-axis (default: "Avg. enrichment")}

\item{plot_title}{The title of the plot (default: "Enrichment per position")}

\item{bar_outline}{The color of the bar outlines (default: "black")}

\item{bar_fill}{The color of the bar fills (default: "skyblue")}
}
\value{
A plotly object with a bar plot showing the average enrichment per position after aggregating the residue-specific enrichment values
}
\description{
\code{fa_posEnrich_barplot} returns a bar plot with one bar per position in the aligned sequences. The height of each bar corresponds to the average enrichment of sequences.
A much more informative plot is generated with \code{fa_posEnrich_heatmap}, so we recommend using that version instead.
}
\examples{
\dontrun{
# count FASTQ files, cluster data, recluster populations together, compute position enrichment
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
clusterData <- list(
  fa_clusterLED(faDF = countData[[1]]),
  fa_clusterLED(faDF = countData[[2]])
)
reclusterData <- fa_recluster(
  faDF1_cluster = clusterData[[1]],
  faDF2_cluster = clusterData[[2]]
)
posEnrichData <- fa_posEnrich(faDF_recluster = reclusterData, cluster_selection = 1)

# generate bar plot
fa_posEnrich_barplot(faDF_posEnrich = posEnrichData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_posEnrich]{fa_posEnrich()}}, \code{\link[=fa_posEnrich_heatmap]{fa_posEnrich_heatmap()}}
}
