% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posEnrich_plots.R
\name{fa_posEnrich_heatmap}
\alias{fa_posEnrich_heatmap}
\title{Return a heat map showing the average enrichment per position, with regard to the character at that position}
\usage{
fa_posEnrich_heatmap(
  faDF_posEnrich = NULL,
  xaxis = "Position",
  yaxis = "Residue",
  legend_title = "Avg. enrichment",
  plot_title = "Enrichment per position per residue",
  fill_palette_cont = "magma"
)
}
\arguments{
\item{faDF_posEnrich}{A data.frame from \code{fa_posEnrich}}

\item{xaxis}{The title of the x-axis (default: "Position")}

\item{yaxis}{The title of the y-axis (default: "Residue")}

\item{legend_title}{The title of the legend (default: "Avg. enrichment")}

\item{plot_title}{The title of the plot (default: "Enrichment per position per residue")}

\item{fill_palette_cont}{A continuous color palette (default: "magma")}
}
\value{
A plotly object with a heat map showing the average enrichment per position per residue at that position
}
\description{
\code{fa_posEnrich_heatmap} returns a heat map in which the x-axis corresponds to each position of an MSA,
the y-axis corresponds to possible characters (e.g., D/RNA populations will only have rows corresponding to the four nucleotides or a gap),
and the color corresponds to the enrichment value of sequences with the indicated character at the indicated position.
}
\examples{
\dontrun{
# count FASTQ files, cluster data, recluster populations together, compute position enrichment
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
clusterData <- list(
  fa_clusterLED(faDF = countData[[1]]),
  fa_clusterLED(faDF = countData[[2]])
)
reclusterData <- fa_recluster(
  faDF1_cluster = clusterData[[1]],
  faDF2_cluster = clusterData[[2]]
)
posEnrichData <- fa_posEnrich(faDF_recluster = reclusterData, cluster_selection = 1)

# make heat map
fa_posEnrich_heatmap(faDF_posEnrich = posEnrichData)
}

}
\seealso{
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}}, \code{\link[=fa_posEnrich]{fa_posEnrich()}}, \code{\link[=fa_posEnrich_barplot]{fa_posEnrich_barplot()}}
}
