% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_analyses.R
\name{fa_preprocess}
\alias{fa_preprocess}
\title{Preprocess FASTA / FASTQ files before counting them}
\usage{
fa_preprocess(
  inputPath = NULL,
  const5p = "",
  const3p = "",
  length_range = c(10, 100),
  max_error = 0.005
)
}
\arguments{
\item{inputPath}{The file path to the FASTA / FASTQ file to preprocess; can be gzipped}

\item{const5p}{A character string corresponding to the 5' constant region to trim (default: "", implying no 5' constant region)}

\item{const3p}{A character string corresponding to the 3' constant region to trim (default: "", implying no 5' constant region)}

\item{length_range}{An integer vector corresponding to the shortest and longest allowable sequences after trimming constant regions (default: 10, 100)}

\item{max_error}{The maximum allowed error in Phred quality (default: 0.005)}
}
\value{
A data.frame with preprocessed sequences; this still needs to pass through \code{fa_count} before it can be used in other FASTAptameR modules
}
\description{
\code{fa_preprocess} 1) trims user-defined constant regions off sequences, 2) omits sequences with a length outside a user-defined range, and 3) omits sequences with an average error above a user-defined threshold.
Trimming occurs with a fuzzy match algorithm that searches for the constant 5' and/or 3' subsequences at the start or end of the sequences, respectively.
The error filter is based on the average Phred quality score for the entire sequence.
}
\examples{
\dontrun{
# preprocess data
preprocessData <- fa_preprocess(
  inputPath = "PATH/TO/FILE.fastq",
  const5p = const5p,
  const3p = const3p,
  length_range = c(50, 60),
  max_error = 0.01
)
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}, \code{\link[utils:aregexec]{utils::aregexec()}}, \code{\link[Biostrings:XStringQuality-class]{Biostrings::PhredQuality()}}
}
