% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster_plots.R
\name{fa_recluster_enrich}
\alias{fa_recluster_enrich}
\title{Return a box plot of the sequence enrichment per cluster}
\usage{
fa_recluster_enrich(
  recluster_df = NULL,
  xaxis = "Cluster",
  plot_title = "Sequence enrichment per super-cluster",
  box_outline = "black",
  box_fill = "skyblue"
)
}
\arguments{
\item{recluster_df}{A reclustered FASTAptameR data.frame from \code{fa_recluster}}

\item{xaxis}{The title of the x-axis (default: "Cluster")}

\item{plot_title}{The title of the plot (default: "Sequence enrichment per super-cluster")}

\item{box_outline}{The color of the box outlines (default: "black")}

\item{box_fill}{The color of the box fills (default: "skyblue")}
}
\value{
A plotly object with a box plot showing the sequence enrichment for each cluster
}
\description{
\code{fa_recluster_enrich} returns a box plot that shows sequence enrichment statistics for each cluster.
}
\examples{
\dontrun{
# count FASTQ files, cluster data, recluster populations together
countData <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)
clusterData <- list(
  fa_clusterLED(faDF = countData[[1]]),
  fa_clusterLED(faDF = countData[[2]])
)
reclusterData <- fa_recluster(
  faDF1_cluster = clusterData[[1]],
  faDF2_cluster = clusterData[[2]]
)

# generate box plot
fa_recluster_enrich(recluster_df = reclusterData)
}

}
\seealso{
\code{\link[grDevices:colors]{grDevices::colors()}}, \code{\link[=fa_recluster]{fa_recluster()}}, \code{\link[=fa_recluster_popSize]{fa_recluster_popSize()}}, \code{\link[=fa_recluster_RPU]{fa_recluster_RPU()}}, \code{\link[=fa_recluster_LED]{fa_recluster_LED()}}
}
