% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_analyses.R
\name{fa_recount}
\alias{fa_recount}
\title{Concatenate sequences from two counted data.frames}
\usage{
fa_recount(faDF1 = NULL, faDF2 = NULL, scaling_factor = 1e+06)
}
\arguments{
\item{faDF1}{A FASTAptameR data.frame}

\item{faDF2}{A FASTAptameR data.frame}

\item{scaling_factor}{The scaling factor to be used for normalizing the read count (default: 1e6)}
}
\value{
A FASTAptameR data.frame
}
\description{
\code{fa_recount} concatenates two counted data.frames, which can be useful when a single file is too large for a single run but can be split into smaller files.
After merging by sequence, this function adds the read count and then renormalizes it based on a common scaling factor.
}
\examples{
\dontrun{
# count FASTQ files
countDataList <- list(
  fa_count(dataInput = "PATH/TO/FILE1.fastq"),
  fa_count(dataInput = "PATH/TO/FILE2.fastq")
)

# concatenate files and recompute sequence features
recountData <- fa_recount(faDF1 = countDataList[[1]], faDF2 = countDataList[[2]])
}

}
\seealso{
\code{\link[=fa_count]{fa_count()}}
}
