% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_analyses.R
\name{fa_translate_mapping}
\alias{fa_translate_mapping}
\title{Format the translation mapping for \code{fa_translate}}
\usage{
fa_translate_mapping(inputChanges = NULL, translateSelection = "Standard")
}
\arguments{
\item{inputChanges}{A data.frame with two columns (Codon and Translation), indicating modifications to the specificed genetic code}

\item{translateSelection}{A standard name that refers to a specific JSON file with codon usage information}
}
\value{
A dictionary-like data.frame to use for translating sequences
}
\description{
\code{fa_translate_mapping} is initialized with a standard genetic code and can be modified with a user-defined data.frame of changes to said code.
The columns of this data.frame represent the three-letter codon and the corresponding one-letter amino acid.
The standard genetic codes are specified in \code{fa_translate}.
Importantly, this is a helper function that is not intended to be used outside of \code{fa_translate}.
}
\seealso{
\code{\link[=fa_translate]{fa_translate()}}
}
