% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_calc_freq}
\alias{fsbc_calc_freq}
\title{Calculate frequency of sequence. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_calc_freq(x, ...)
}
\arguments{
\item{x}{DNAStringset, sequence dataset.}

\item{...}{do not use.}
}
\value{
DNAStringSet, unique sequence dataset with frequency.
}
\description{
The frequency, ratio, and ranking are stored in metadata in the DNAStringSet object.
}
