% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_calc_prob_seq}
\alias{fsbc_calc_prob_seq}
\title{Calculation of probability of sequences including specific string. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_calc_prob_seq(x, symbols, L = 40, ...)
}
\arguments{
\item{x}{string.}

\item{symbols}{probability of symbols.}

\item{L}{length of sequence.}

\item{...}{do not use.}
}
\value{
probability of a sequence including specific string
}
\description{
This function calculates the probability of $L$-mer sequence which includes $l$-mer string.
For example, the probabilit of 5-mer sequence which includes "ATGC" (The probabilities of A, T, G, and C are 1/4.) is 0.0781250.
}
