% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_calc_prob_subseq}
\alias{fsbc_calc_prob_subseq}
\title{Calculate the probability of string. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_calc_prob_subseq(x, symbols, ...)
}
\arguments{
\item{x}{string.}

\item{symbols}{probability of symbols.}

\item{...}{do not use.}
}
\value{
probability of string.
}
\description{
If all nucleobases are equal to 1/4, the probability of "ATGC" is (1/4)^4 = 0.00390625.
}
