% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_calc_subseq_score}
\alias{fsbc_calc_subseq_score}
\title{Calculate Z-score, of string. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).
This function calculates Z-score with sequence probability and observed frequency.
Z-score is provaded with frequency, ratio, and probability of sequences which include string.}
\usage{
fsbc_calc_subseq_score(x, y, symbols, freq, L = width(y), fixed = TRUE, ...)
}
\arguments{
\item{x}{string.}

\item{y}{DNAStringSet, sequence dataset.}

\item{symbols}{probability of symbols.}

\item{freq}{frequency of \code{y}.}

\item{L}{length of sequence.}

\item{fixed}{parameter for vcountPattern.}

\item{...}{do not use.}
}
\value{
results of subseq scores
}
\description{
Calculate Z-score, of string. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).
This function calculates Z-score with sequence probability and observed frequency.
Z-score is provaded with frequency, ratio, and probability of sequences which include string.
}
