% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_check_self_overlap}
\alias{fsbc_check_self_overlap}
\title{Check self overlapping regions of the string. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_check_self_overlap(x, ...)
}
\arguments{
\item{x}{string.}

\item{...}{do not use.}
}
\value{
the set of self overlapping region of string.
}
\description{
If the string is "ATATA", the set of self overlapping is "A" and "ATA".
The following is an example of string "ATATA".
Self overlapping with "A".
ATATA
ATATA
}
\details{
Self overlapping with "ATA".
ATATA
ATATA
}
