% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_extend_subseq}
\alias{fsbc_extend_subseq}
\title{Extend string with each symbol. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_extend_subseq(x, symbols, ...)
}
\arguments{
\item{x}{data.frame of string information.}

\item{symbols}{probabity of symbols.}

\item{...}{do not use.}
}
\value{
extended strings
}
\description{
This function returns the extended strings.
If the string is "ATGC", this function returns {"ATGCA", "ATGCT", "ATGCC", "ATGCG"}.
}
