% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_get_base_ratio}
\alias{fsbc_get_base_ratio}
\title{Get ratio of nucleobases. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_get_base_ratio(x, symbols = DNA_BASES, ...)
}
\arguments{
\item{x}{DNAStringSet of sequence data.}

\item{symbols}{DNA_BASES.}

\item{...}{do not use.}
}
\value{
ratio of nucleobases.
}
\description{
This function calculates the ratio of nucleobases with all sequence data.
}
