% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_get_initial_subseq}
\alias{fsbc_get_initial_subseq}
\title{Get the set of initial strings. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_get_initial_subseq(symbols, l, r = l, ...)
}
\arguments{
\item{symbols}{symbols (symbols are nucleobases).}

\item{l}{length of string.}

\item{r}{length of string.}

\item{...}{do not use.}
}
\value{
set of initial strings.
}
\description{
This function generates all $l$-mer strings.
If $l$ is 2, this function generates
{ AA, AT, AG, AC, TA, TT, TG, TG, GA, GT, GG, GC, CA, CT, CG, CC }.
}
