% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_label_cluster}
\alias{fsbc_label_cluster}
\title{Add cluster IDs to names of DNAStringSet dataset. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_label_cluster(x, y, ...)
}
\arguments{
\item{x}{List of the clusters.}

\item{y}{DNSStringSet, sequence datset.}

\item{...}{do not use}
}
\value{
labeled clusters
}
\description{
This function adds cluster ID to the DNAStringSet dataset, and also add cluster ID and cluster seeds of subsequece to \code{y}.
}
