% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_perm}
\alias{fsbc_perm}
\title{Get permutations. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_perm(x, n, res = "", ...)
}
\arguments{
\item{x}{vector of symbols.}

\item{n}{the number of permutations.}

\item{res}{characters}

\item{...}{do not use.}
}
\description{
Get permutations. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).
}
