% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_scale}
\alias{fsbc_scale}
\title{This is a wrapper function of "scale" to avoid errors of NULL. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_scale(x, ...)
}
\arguments{
\item{x}{numerical vector.}

\item{...}{parameters for scale function.}
}
\value{
results of scale.
}
\description{
Scale function shows error with NULL.
This function return NULL, if the value is NULL for avoiding error of scale function.
}
