% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_search_subseq}
\alias{fsbc_search_subseq}
\title{Select strings with eliminating the search space. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_search_subseq(
  x,
  freq,
  lmin = 5,
  lmax = 10,
  symbols = fsbc_get_base_ratio(x, symbols = DNA_BASES),
  s0 = fsbc_get_initial_subseq(names(symbols), lmin),
  fun = function(x, y) x > y,
  ...
)
}
\arguments{
\item{x}{DNAStringSet, non-redundant sequence file.}

\item{freq}{frequency of sequences of \code{x}. This have to be same order with DNAStringSet \code{x}.}

\item{lmin}{minimum length of string.}

\item{lmax}{maximum length of string.}

\item{symbols}{probability of symbols.}

\item{s0}{initial string.}

\item{fun}{carry over criteria for branch and bound.}

\item{...}{functions for fsbc_calc_subseq_score function.}
}
\value{
data.frame of selected strings.
}
\description{
This function select strings with lengths from $l$-min to $l$-max with eliminating the search space.
strings of $l$-min length are all enumerated for initial subset of string searching.
The elimination rule of function is $x > y$, where $x$ and $y$ are pre and post extended string of Z-scores.
The elimination rule is flexible to change with the parameter \code{fun}.
}
