% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_FSBC.R
\name{fsbc_seq_cluster}
\alias{fsbc_seq_cluster}
\title{Generate clusters based on string information. FSBC functions are from "FSBC: FAST STRING-BASED CLUSTERING FOR HT-SELEX DATA" by Kato et al. (DOI: 10.1186/s12859-020-03607-1).}
\usage{
fsbc_seq_cluster(x, y, res = list(), ...)
}
\arguments{
\item{x}{vector of strings.}

\item{y}{DNAStringSet, sequence dataset.}

\item{res}{result object for this recursive function. Default is list().}

\item{...}{do not use}
}
\value{
list of sequence clusters
}
\description{
This function generates clusters with ordered strings, recursively.
}
